Bugfix: Only instances which locked can unlock

+++ Program/src/org/jphototagger/program/app/AppStartupLock.java
@@ -14,17 +14,18 @@ import org.jphototagger.lib.util.Bundle;
 
 /**
  * Creates an application lock file to prevent multiple instances.
+ *
  * @author Elmar Baumann
  */
 public final class AppStartupLock {
 
     private static final String LOCKFILE_NAME;
+    private static boolean hasUnlockPrivilege;
 
     static {
         FileRepositoryProvider provider = Lookup.getDefault().lookup(FileRepositoryProvider.class);
         File repositoryDirectory = provider.getFileRepositoryDirectory();
         String repositoryDirectoryName = repositoryDirectory.getAbsolutePath();
-
         LOCKFILE_NAME = repositoryDirectoryName + File.separator + AppInfo.PROJECT_NAME + ".lck";
         addShutdownCleanupCheck();
     }
@@ -40,34 +41,31 @@ public final class AppStartupLock {
         if (!isLocked()) {
             try {
                 FileUtil.ensureFileExists(new File(LOCKFILE_NAME));
-
+                hasUnlockPrivilege = true;
                 return true;
             } catch (IOException ex) {
                 Logger.getLogger(AppStartupLock.class.getName()).log(Level.SEVERE, null, ex);
             }
         }
-
         return false;
     }
 
     public static synchronized boolean unlock() {
-        if (isLocked()) {
+        if (hasUnlockPrivilege && isLocked()) {
             return new File(LOCKFILE_NAME).delete();
         }
-
         return true;
     }
 
     /**
-     * Displays an error message that the lock file couldn't be created and
-     * offers the option to unlock - delete - it.
+     * Displays an error message that the lock file couldn't be created and offers the option to unlock - delete - it.
+     *
      * @return true if unlocked
      */
     public static synchronized boolean forceLock() {
         if (confirmForceUnlock()) {
             return deleteLockFile() && lock();
         }
-
         return false;
     }
 
@@ -75,25 +73,21 @@ public final class AppStartupLock {
         if (!FileUtil.existsFile(new File(LOCKFILE_NAME))) {
             return true;
         }
-
         if (new File(LOCKFILE_NAME).delete()) {
             return true;
         } else {
             errorMessageDelete();
-
             return false;
         }
     }
 
     private static boolean confirmForceUnlock() {
         String message = Bundle.getString(AppStartupLock.class, "AppStartupLock.Error.LockFileExists", LOCKFILE_NAME);
-
         return MessageDisplayer.confirmYesNo(null, message);
     }
 
     private static void errorMessageDelete() {
         String message = Bundle.getString(AppStartupLock.class, "AppStartupLock.Error.DeleteLockFile", LOCKFILE_NAME);
-
         MessageDisplayer.error(null, message);
     }
 
@@ -101,9 +95,7 @@ public final class AppStartupLock {
         Runtime.getRuntime().addShutdownHook(new Thread() {
 
             @Override public void run() {
-                if (new File(LOCKFILE_NAME).exists()) {
-                    unlock();
-                }
+                unlock();
             }
         });
     }
